BugFix:: on methode "template <class MEMORY_ADDR> bool DWARF_ExpressionVM<MEMORY_ADDR>::Run()" at   - case 2 : replace uint8_t by uint16_t   - case 4 : replace uint8_t by uint32_t
Former-commit-id: baa02f41becda996f4aaa26d7a3cca97d605356d

+++ unisim_lib/unisim/service/loader/elf_loader/dwarf.tcc
@@ -2829,7 +2829,8 @@ bool DWARF_ExpressionVM<MEMORY_ADDR>::Run(const DWARF_Expression<MEMORY_ADDR> *d
 								break;
 							case 2:
 								{
-									uint8_t addr16;
+									// *** REDA *** replace uint8_t by uint16_t
+									uint16_t addr16;
 									memcpy(&addr16, expr, address_size);
 									addr16 = Target2Host(endianness, addr16);
 									addr = addr16;
@@ -2837,7 +2838,8 @@ bool DWARF_ExpressionVM<MEMORY_ADDR>::Run(const DWARF_Expression<MEMORY_ADDR> *d
 								break;
 							case 4:
 								{
-									uint8_t addr32;
+									// *** REDA *** replace uint8_t by uint32_t
+									uint32_t addr32;
 									memcpy(&addr32, expr, address_size);
 									addr32 = Target2Host(endianness, addr32);
 									addr = addr32;
